+++ serial-motor-control/structures.h
@@ -2,7 +2,7 @@ typedef struct {
   int res;
   int dir;
   int steps;
-  int stepDelay;
+  unsigned long stepDelay;
   int ref;
   int inProgress;
   unsigned long previousMicros;
